+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Ensure no tasks are scheduled prior to 10:00 or during any meetings
 * Add the amount of tasks to the daily log output
 * Add labels to daily log task list
+* Add todoist backup and sync
 
 # 1.1.6: 2024-10-20